+++ common/src/main/java/io/github/mortuusars/exposure/client/capture/template/PathCaptureTemplate.java
@@ -44,7 +44,7 @@ public class PathCaptureTemplate implements CaptureTemplate {
         return Capture.of(Capture.path(path),
                         CaptureAction.optional(params.cameraId(), CaptureAction::interplanarProjection))
                 .logErrorAndGetResult(LOGGER)
-                .thenAsync(applyEffectsToImage(params))
+                .thenAsync(applyEffectsToImage(params.mutable().setCropFactor(1f).build())) // Remove crop factor for loaded images
                 .thenAsync(Palettizer.fromProjectionMode(projection.mode()).palettizeAndClose(palette.value()))
                 .thenAsync(convertToExposureData(palette, createExposureTag(params, true)))
                 .acceptAsync(image -> ExposureUploader.upload(params.exposureId(), image))